requirements.txt
pandas
streamlit
scikit-surprise
